// client/src/app/stores/jobs/job.types.ts
export enum JOBS_ACTION_TYPES {
    FETCH_ALL_JOBS_LOADING = "jobs/FETCH_ALL_JOBS_LOADING",
    FETCH_ALL_JOBS_SUCCESS = "jobs/FETCH_ALL_JOBS_SUCCESS",
    FETCH_ALL_JOBS_FAILED = "jobs/FETCH_ALL_JOBS_FAILED",
    FETCH_JOB_BY_ID_LOADING = "jobs/FETCH_JOB_BY_ID_LOADING",
    FETCH_JOB_BY_ID_SUCCESS = "jobs/FETCH_JOB_BY_ID_SUCCESS",
    FETCH_JOB_BY_ID_FAILED = "jobs/FETCH_JOB_BY_ID_FAILED",
    ADD_JOB_LOADING = "jobs/ADD_JOB_LOADING",
    ADD_JOB_SUCCESS = "jobs/ADD_JOB_SUCCESS",
    ADD_JOB_FAILED = "jobs/ADD_JOB_FAILED",
    UPDATE_JOB_LOADING = "jobs/UPDATE_JOB_LOADING",
    UPDATE_JOB_SUCCESS = "jobs/UPDATE_JOB_SUCCESS",
    UPDATE_JOB_FAILED = "jobs/UPDATE_JOB_FAILED",
    DELETE_JOB_LOADING = "jobs/DELETE_JOB_LOADING",
    DELETE_JOB_SUCCESS = "jobs/DELETE_JOB_SUCCESS",
    DELETE_JOB_FAILED = "jobs/DELETE_JOB_FAILED",
  };